drivers: mmidp: Fix incorrect datatype of PixelClockHz
The setting of certain modes fail due to incorrect datatype of
PixelClockHz. This update fixes and supports all the required pixel
clock frequency values.


Acked-for-series: Varunkumar Allagadapa <varunkumar.allagadapa@amd.com>

+++ XilinxProcessorIPLib/drivers/mmidp/src/xmmidp.h
@@ -297,7 +297,7 @@ typedef struct {
 	u32 NVid;
 	u32 HBlankInterval;
 	u8 BitsPerColor;
-	u8 PixelClockHz;
+	u64 PixelClockHz;
 	u8 Reserved[2];
 } XMmiDp_MainStreamAttributes;
